+++ src/Onboarding/Setup/templates/index.html.php
@@ -12,10 +12,6 @@
 	</div>
 	<?php endif; ?>
 
-	<p class="intro-text">
-		<?php echo __( 'You\'re almost ready to start fundraising for your cause! Below we’ll guide you through the final steps to get up and running quickly.', 'give' ); ?>
-	</p>
-
 	<!-- Configuration -->
 	<?php
 		echo $this->render_template(
@@ -62,7 +58,7 @@
 												: $this->image( 'paypal@2x.min.png' ),
 							'icon_alt'    => esc_html__( 'PayPal', 'give' ),
 							'title'       => esc_html__( 'Connect to PayPal', 'give' ),
-							'description' => esc_html__( 'PayPal is synonymous with nonprofits and online charitable gifts. It\’s been the go-to payment merchant for many of the world\'s top NGOs. Accept PayPal, credit and debit cards without any added platform fees.', 'give' ),
+							'description' => esc_html__( 'PayPal is synonymous with nonprofits and online charitable gifts. It\'s been the go-to payment merchant for many of the world\'s top NGOs. Accept PayPal, credit and debit cards without any added platform fees.', 'give' ),
 							'action'      => sprintf(
 								'<a href="%s"><i class="fab fa-paypal"></i>&nbsp;&nbsp;Connect to PayPal</a>',
 								add_query_arg(