+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- Changes made since the last release are stored here until a release is ready. -->
 
 ### Changed
+
+-   Update Onboarding copy text and typos. (#5168)
 -   Reduced padding on the Setup Page container. (#5165)
 -   Normalizes format for country and state selects, with states having an empty option and countries not. (#5163)
 -   Reduced the scope of marked optional fields in the Multi-Step template to the User Info fieldset. (#5161)